Перенести компонент App в components/
