+++ src/components/app/app.jsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import WelcomeScreen from '../welcome-screen/welcome-screen.jsx';
+
+/**
+ * @description Компонент приложения
+ * @param {Object} props игорвое время
+ * @param {Number} props.gameTime игорвое время
+ * @param {Number} props.errorCount допустимое количество ошибок
+ * @return {WelcomeScreen}
+ */
+const App = (props) => {
+  const {time: gameTime, errorCount} = props;
+
+  return <WelcomeScreen
+    time={gameTime}
+    errorCount={errorCount}
+  />;
+};
+
+App.propTypes = {
+  time: PropTypes.number.isRequired,
+  errorCount: PropTypes.number.isRequired
+};
+
+export default App;
